EES-1228 Replacing DataBlockRequest with ObservationQueryContext since they should be the same, and refactoring IContentBlockChart to be shared by Admin and Publisher view models as well as the Content model.
