+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Controllers/Api/Statistics/DataController.cs
@@ -1,5 +1,6 @@
 using System.Diagnostics;
 using System.Web.Http;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
 using GovUk.Education.ExploreEducationStatistics.Data.Model.Query;
 using GovUk.Education.ExploreEducationStatistics.Data.Services.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Data.Services.ViewModels;
