+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Mappings/DataMappingProfiles.cs
@@ -4,6 +4,8 @@ using System.Linq;
 using System.Reflection;
 using AutoMapper;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
+using GovUk.Education.ExploreEducationStatistics.Common.Model;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data;
 using GovUk.Education.ExploreEducationStatistics.Data.Model;
 using GovUk.Education.ExploreEducationStatistics.Data.Services.ViewModels;
 using GovUk.Education.ExploreEducationStatistics.Data.Services.ViewModels.Meta;
