+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Models/Api/CreateDataBlockViewModel.cs
@@ -1,4 +1,6 @@
 using System.Collections.Generic;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Chart;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
 using GovUk.Education.ExploreEducationStatistics.Content.Model;
 
 namespace GovUk.Education.ExploreEducationStatistics.Admin.Models.Api
@@ -6,14 +8,14 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Models.Api
     public class CreateDataBlockViewModel
     {
         public string Heading { get; set; }
-        
+
         public string CustomFootnotes { get; set; }
-        
+
         public string Name { get; set; }
-        
+
         public string Source { get; set; }
 
-        public DataBlockRequest DataBlockRequest { get; set; }
+        public ObservationQueryContext DataBlockRequest { get; set; }
 
         public List<IContentBlockChart> Charts { get; set; }
