+++ src/GovUk.Education.ExploreEducationStatistics.Common.Tests/Converters/ContentBlockChartConverterTests.cs
@@ -1,9 +1,10 @@
-using GovUk.Education.ExploreEducationStatistics.Content.Model.Converters;
+using GovUk.Education.ExploreEducationStatistics.Common.Converters;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Chart;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using Xunit;
 
-namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Tests.Converters
+namespace GovUk.Education.ExploreEducationStatistics.Common.Tests.Converters
 {
     public class ContentBlockChartConverterTests
     {
@@ -11,7 +12,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Tests.Convert
         public void ContentBlockChartConverterReturnsLine()
         {
             const string testString = "{\"Type\":\"line\"}";
-            
+
             var serializerSettings = new JsonSerializerSettings();
             serializerSettings.Converters.Add(new ContentBlockChartConverter());
 
@@ -19,15 +20,15 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Tests.Convert
 
             var jObject = JObject.Parse(testString);
             var result = jObject.ToObject<LineChart>(serializer);
-            
+
             Assert.IsType<LineChart>(result);
         }
-        
+
         [Fact]
         public void ContentBlockChartConverterReturnsHorizontalBar()
         {
             const string testString = "{\"Type\":\"horizontalbar\"}";
-            
+
             var serializerSettings = new JsonSerializerSettings();
             serializerSettings.Converters.Add(new ContentBlockChartConverter());
 
@@ -35,15 +36,15 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Tests.Convert
 
             var jObject = JObject.Parse(testString);
             var result = jObject.ToObject<HorizontalBarChart>(serializer);
-            
+
             Assert.IsType<HorizontalBarChart>(result);
         }
-        
+
         [Fact]
         public void ContentBlockChartConverterReturnsVerticalBar()
         {
             const string testString = "{\"Type\":\"verticalbar\"}";
-            
+
             var serializerSettings = new JsonSerializerSettings();
             serializerSettings.Converters.Add(new ContentBlockChartConverter());
 
@@ -51,15 +52,15 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Tests.Convert
 
             var jObject = JObject.Parse(testString);
             var result = jObject.ToObject<VerticalBarChart>(serializer);
-            
+
             Assert.IsType<VerticalBarChart>(result);
         }
-        
+
         [Fact]
         public void ContentBlockChartConverterReturnsMap()
         {
             const string testString = "{\"Type\":\"map\"}";
-            
+
             var serializerSettings = new JsonSerializerSettings();
             serializerSettings.Converters.Add(new ContentBlockChartConverter());
 
@@ -67,7 +68,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Tests.Convert
 
             var jObject = JObject.Parse(testString);
             var result = jObject.ToObject<MapChart>(serializer);
-            
+
             Assert.IsType<MapChart>(result);
         }
     }
