+++ src/GovUk.Education.ExploreEducationStatistics.Common/Model/Chart/ChartDataLocation.cs
@@ -1,0 +1,19 @@
+﻿namespace GovUk.Education.ExploreEducationStatistics.Common.Model.Chart
+{
+    public class ChartDataLocation
+    {
+        public ChartDataLocationValue Country;
+        public ChartDataLocationValue Institution;
+        public ChartDataLocationValue LocalAuthority;
+        public ChartDataLocationValue LocalAuthorityDistrict;
+        public ChartDataLocationValue LocalEnterprisePartnership;
+        public ChartDataLocationValue MultiAcademyTrust;
+        public ChartDataLocationValue MayoralCombinedAuthority;
+        public ChartDataLocationValue OpportunityArea;
+        public ChartDataLocationValue ParliamentaryConstituency;
+        public ChartDataLocationValue Region;
+        public ChartDataLocationValue RscRegion;
+        public ChartDataLocationValue Sponsor;
+        public ChartDataLocationValue Ward;
+    }
+}
