+++ src/GovUk.Education.ExploreEducationStatistics.Common/Model/Chart/ChartDataLocationValue.cs
@@ -1,0 +1,8 @@
+﻿namespace GovUk.Education.ExploreEducationStatistics.Common.Model.Chart
+{
+    public class ChartDataLocationValue
+    {
+        public string Code { get; set; }
+        public string Name { get; set; }
+    }
+}
