+++ src/GovUk.Education.ExploreEducationStatistics.Common/Model/Data/BoundaryLevel.cs
@@ -1,7 +1,6 @@
 using System;
-using GovUk.Education.ExploreEducationStatistics.Common.Model;
 
-namespace GovUk.Education.ExploreEducationStatistics.Data.Model
+namespace GovUk.Education.ExploreEducationStatistics.Common.Model.Data
 {
     public class BoundaryLevel
     {
