+++ src/GovUk.Education.ExploreEducationStatistics.Common/Model/Data/GeographicLevel.cs
@@ -1,6 +1,6 @@
 using GovUk.Education.ExploreEducationStatistics.Common.Database;
 
-namespace GovUk.Education.ExploreEducationStatistics.Common.Model
+namespace GovUk.Education.ExploreEducationStatistics.Common.Model.Data
 {
     public enum GeographicLevel
     {
