+++ src/GovUk.Education.ExploreEducationStatistics.Common/Model/Data/Query/TimePeriodQuery.cs
@@ -1,9 +1,8 @@
 using GovUk.Education.ExploreEducationStatistics.Common.Converters;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
-using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using Newtonsoft.Json;
 
-namespace GovUk.Education.ExploreEducationStatistics.Data.Model.Query
+namespace GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query
 {
     public class TimePeriodQuery
     {
