+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model/Database/ContentDbContext.cs
@@ -5,6 +5,9 @@ using System.Text;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using GovUk.Education.ExploreEducationStatistics.Common.Converters;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Chart;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Newtonsoft.Json;
@@ -308,7 +311,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                 .HasColumnName("DataBlock_Request")
                 .HasConversion(
                     v => JsonConvert.SerializeObject(v),
-                    v => JsonConvert.DeserializeObject<DataBlockRequest>(v));
+                    v => JsonConvert.DeserializeObject<ObservationQueryContext>(v));
 
             modelBuilder.Entity<DataBlock>()
                 .Property(block => block.Charts)
@@ -2500,18 +2503,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("7b779d79-6caa-43fd-84ba-b8efd219b3c8"),
                     Order = 1,
                     Name = "Key Stat 1",
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.AbsenceByCharacteristic],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
                         Filters = new List<Guid>
@@ -2580,9 +2583,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -2626,7 +2629,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Absence Rate"
@@ -2654,18 +2657,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("7b779d79-6caa-43fd-84ba-b8efd219b3c8"),
                     Name = "Key Stat 2",
                     Order = 2,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.AbsenceByCharacteristic],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
                         Filters = new List<Guid>
@@ -2734,9 +2737,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -2756,7 +2759,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Absence Rate"
@@ -2784,18 +2787,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("7b779d79-6caa-43fd-84ba-b8efd219b3c8"),
                     Name = "Key Stat 3",
                     Order = 3,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.AbsenceByCharacteristic],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
                         Filters = new List<Guid>
@@ -2864,9 +2867,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -2910,7 +2913,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Absence Rate"
@@ -2938,18 +2941,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("30d74065-66b8-4843-9761-4578519e1394"),
                     Name = "Key Stats aggregate table",
                     Order = 1,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.AbsenceByCharacteristic],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
                         Filters = new List<Guid>
@@ -3034,9 +3037,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -3080,7 +3083,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Absence Rate"
@@ -3123,18 +3126,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     Id = new Guid("5d3058f2-459e-426a-b0b3-9f60d8629fef"),
                     ContentSectionId = new Guid("8965ef44-5ad7-4ab0-a142-78453d6f40af"),
                     Name = "Generic data block - National",
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.AbsenceByCharacteristic],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
                         Filters = new List<Guid>
@@ -3188,9 +3191,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -3234,7 +3237,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Absence Rate"
@@ -3276,15 +3279,15 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     Id = new Guid("4a1af98a-ed8a-438e-92d4-d21cca0429f9"),
                     ContentSectionId = new Guid("68e3028c-1291-42b3-9e7c-9be285dac9a1"),
                     Name = "Generic data block - LA",
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.AbsenceByCharacteristic],
                         GeographicLevel = GeographicLevel.LocalAuthorityDistrict,
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2016",
+                            StartYear = 2016,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2017",
+                            EndYear = 2017,
                             EndCode = TimeIdentifier.AcademicYear
                         },
 
@@ -3308,9 +3311,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new MapChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -3354,7 +3357,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Title = "Absence Rate"
                                 }
@@ -3397,18 +3400,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("991a436a-9c7a-418b-ab06-60f2610b4bc6"),
                     Name = "Key Stat 1",
                     Order = 1,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.ExclusionsByGeographicLevel],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
 
@@ -3477,9 +3480,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -3507,7 +3510,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Absence Rate"
@@ -3537,18 +3540,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("991a436a-9c7a-418b-ab06-60f2610b4bc6"),
                     Name = "Key Stat 2",
                     Order = 2,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.ExclusionsByGeographicLevel],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
 
@@ -3616,9 +3619,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -3636,7 +3639,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Absence Rate"
@@ -3666,18 +3669,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("991a436a-9c7a-418b-ab06-60f2610b4bc6"),
                     Name = "Key Stat 3",
                     Order = 3,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.ExclusionsByGeographicLevel],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
 
@@ -3746,9 +3749,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -3766,7 +3769,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Absence Rate"
@@ -3796,18 +3799,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("e8a813ce-c68a-417b-af31-91db19377b10"),
                     Name = "Key Stats aggregate table",
                     Order = 1,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.ExclusionsByGeographicLevel],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
 
@@ -3900,9 +3903,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -3930,7 +3933,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Absence Rate"
@@ -3968,18 +3971,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("6ed87fd1-81a5-46dc-8841-4598bdae7fee"),
                     Heading = "Chart showing permanent exclusions in England",
                     Name = "Generic data block 1",
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.ExclusionsByGeographicLevel],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
                         Filters = new List<Guid>
@@ -4032,9 +4035,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -4052,7 +4055,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Exclusion Rate"
@@ -4080,18 +4083,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("7981db34-afdb-4f84-99e8-bfd43e58f16d"),
                     Heading = "Chart showing fixed-period exclusions in England",
                     Name = "Generic data block 2",
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.ExclusionsByGeographicLevel],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
 
@@ -4145,9 +4148,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -4165,7 +4168,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Absence Rate"
@@ -4195,18 +4198,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = null, // not yet used in any Content
                     Name = "Available Data Block",
                     Order = 0,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.ExclusionsByGeographicLevel],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
 
@@ -4275,9 +4278,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -4295,7 +4298,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Absence Rate"
@@ -4325,18 +4328,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = null, // not yet used in any Content
                     Name = "Available Data Block 2",
                     Order = 0,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.ExclusionsByGeographicLevel],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2012",
+                            StartYear = 2012,
                             StartCode = TimeIdentifier.AcademicYear,
-                            EndYear = "2016",
+                            EndYear = 2016,
                             EndCode = TimeIdentifier.AcademicYear
                         },
 
@@ -4405,9 +4408,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     {
                         new LineChart
                         {
-                            Axes = new Dictionary<string, AxisConfigurationItem>
+                            Axes = new Dictionary<string, ChartAxisConfigurationItem>
                             {
-                                ["major"] = new AxisConfigurationItem
+                                ["major"] = new ChartAxisConfigurationItem
                                 {
                                     GroupBy = AxisGroupBy.timePeriod,
                                     DataSets = new List<ChartDataSet>
@@ -4425,7 +4428,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                                     },
                                     Title = "School Year"
                                 },
-                                ["minor"] = new AxisConfigurationItem
+                                ["minor"] = new ChartAxisConfigurationItem
                                 {
                                     Min = 0,
                                     Title = "Absence Rate"
@@ -4455,18 +4458,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("de8f8547-cbae-4d52-88ec-d78d0ad836ae"),
                     Name = "Key Stat 1",
                     Order = 1,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.SchoolApplicationsAndOffers],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2014",
+                            StartYear = 2014,
                             StartCode = TimeIdentifier.CalendarYear,
-                            EndYear = "2018",
+                            EndYear = 2018,
                             EndCode = TimeIdentifier.CalendarYear
                         },
                         Filters = new List<Guid>
@@ -4544,18 +4547,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("de8f8547-cbae-4d52-88ec-d78d0ad836ae"),
                     Name = "Key Stat 2",
                     Order = 2,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.SchoolApplicationsAndOffers],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2014",
+                            StartYear = 2014,
                             StartCode = TimeIdentifier.CalendarYear,
-                            EndYear = "2018",
+                            EndYear = 2018,
                             EndCode = TimeIdentifier.CalendarYear
                         },
                         Filters = new List<Guid>
@@ -4633,18 +4636,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("de8f8547-cbae-4d52-88ec-d78d0ad836ae"),
                     Name = "Key Stat 3",
                     Order = 3,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.SchoolApplicationsAndOffers],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2014",
+                            StartYear = 2014,
                             StartCode = TimeIdentifier.CalendarYear,
-                            EndYear = "2018",
+                            EndYear = 2018,
                             EndCode = TimeIdentifier.CalendarYear
                         },
                         Filters = new List<Guid>
@@ -4722,18 +4725,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     ContentSectionId = new Guid("39c298e9-6c5f-47be-85cb-6e49b1b1931f"),
                     Name = "Key Stats aggregate table",
                     Order = 1,
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.SchoolApplicationsAndOffers],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2014",
+                            StartYear = 2014,
                             StartCode = TimeIdentifier.CalendarYear,
-                            EndYear = "2018",
+                            EndYear = 2018,
                             EndCode = TimeIdentifier.CalendarYear
                         },
                         Filters = new List<Guid>
@@ -4843,18 +4846,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     Heading =
                         "Table of Timeseries of key secondary preference rates, England",
                     Name = "Generic data block 1",
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.SchoolApplicationsAndOffers],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2014",
+                            StartYear = 2014,
                             StartCode = TimeIdentifier.CalendarYear,
-                            EndYear = "2018",
+                            EndYear = 2018,
                             EndCode = TimeIdentifier.CalendarYear
                         },
 
@@ -4920,18 +4923,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Database
                     Heading =
                         "Table showing Timeseries of key primary preference rates, England Entry into academic year",
                     Name = "Generic data block 2",
-                    DataBlockRequest = new DataBlockRequest
+                    DataBlockRequest = new ObservationQueryContext
                     {
                         SubjectId = SubjectIds[SubjectName.SchoolApplicationsAndOffers],
                         Country = new List<string>
                         {
                             CountryCodeEngland
                         },
-                        TimePeriod = new TimePeriod
+                        TimePeriod = new TimePeriodQuery
                         {
-                            StartYear = "2014",
+                            StartYear = 2014,
                             StartCode = TimeIdentifier.CalendarYear,
-                            EndYear = "2018",
+                            EndYear = 2018,
                             EndCode = TimeIdentifier.CalendarYear
                         },
