+++ src/GovUk.Education.ExploreEducationStatistics.Data.Api/Controllers/DataController.cs
@@ -1,4 +1,5 @@
 using System.Diagnostics;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
 using GovUk.Education.ExploreEducationStatistics.Data.Model.Query;
 using GovUk.Education.ExploreEducationStatistics.Data.Services.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Data.Services.ViewModels;
