+++ src/GovUk.Education.ExploreEducationStatistics.Data.Api/Mappings/MappingProfiles.cs
@@ -2,6 +2,8 @@ using System;
 using System.Linq;
 using AutoMapper;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
+using GovUk.Education.ExploreEducationStatistics.Common.Model;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data;
 using GovUk.Education.ExploreEducationStatistics.Data.Api.Models;
 using GovUk.Education.ExploreEducationStatistics.Data.Api.ViewModels;
 using GovUk.Education.ExploreEducationStatistics.Data.Model;
