+++ src/GovUk.Education.ExploreEducationStatistics.Data.Model/Query/SubjectMetaQueryContext.cs
@@ -1,7 +1,8 @@
 using System;
 using System.Collections.Generic;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
-using GovUk.Education.ExploreEducationStatistics.Common.Model;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
 
 namespace GovUk.Education.ExploreEducationStatistics.Data.Model.Query
 {
@@ -25,6 +26,32 @@ namespace GovUk.Education.ExploreEducationStatistics.Data.Model.Query
         public IEnumerable<string> RscRegion { get; set; }
         public IEnumerable<string> Sponsor { get; set; }
         public IEnumerable<string> Ward { get; set; }
+
+        public static SubjectMetaQueryContext FromObservationQueryContext(
+            ObservationQueryContext observationQueryContext)
+        {
+            return new SubjectMetaQueryContext
+            {
+                SubjectId = observationQueryContext.SubjectId,
+                TimePeriod = observationQueryContext.TimePeriod,
+                BoundaryLevel = observationQueryContext.BoundaryLevel,
+                GeographicLevel = observationQueryContext.GeographicLevel,
+                Indicators = observationQueryContext.Indicators,
+                Country = observationQueryContext.Country,
+                Institution = observationQueryContext.Institution,
+                LocalAuthority = observationQueryContext.LocalAuthority,
+                LocalAuthorityDistrict = observationQueryContext.LocalAuthorityDistrict,
+                LocalEnterprisePartnership = observationQueryContext.LocalEnterprisePartnership,
+                MultiAcademyTrust = observationQueryContext.MultiAcademyTrust,
+                MayoralCombinedAuthority = observationQueryContext.MayoralCombinedAuthority,
+                OpportunityArea = observationQueryContext.OpportunityArea,
+                ParliamentaryConstituency = observationQueryContext.ParliamentaryConstituency,
+                Region = observationQueryContext.Region,
+                RscRegion = observationQueryContext.RscRegion,
+                Sponsor = observationQueryContext.Sponsor,
+                Ward = observationQueryContext.Ward
+            };
+        }
 
         public override string ToString()
         {
