+++ src/GovUk.Education.ExploreEducationStatistics.Data.Model/Query/TableBuilderQueryContext.cs
@@ -1,3 +1,5 @@
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
+
 namespace GovUk.Education.ExploreEducationStatistics.Data.Model.Query
 {
     public class TableBuilderQueryContext : ObservationQueryContext
