+++ src/GovUk.Education.ExploreEducationStatistics.Data.Model/Services/Interfaces/IBoundaryLevelService.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data;
 
 namespace GovUk.Education.ExploreEducationStatistics.Data.Model.Services.Interfaces
 {
