+++ src/GovUk.Education.ExploreEducationStatistics.Data.Services.Tests/TimePeriodUtilTests.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
-using GovUk.Education.ExploreEducationStatistics.Data.Model.Query;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Xunit;
 using static GovUk.Education.ExploreEducationStatistics.Common.Model.TimeIdentifier;
