+++ src/GovUk.Education.ExploreEducationStatistics.Data.Services/Interfaces/IDataService.cs
@@ -1,3 +1,4 @@
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
 using GovUk.Education.ExploreEducationStatistics.Data.Model.Query;
 using ReleaseId = System.Guid;
