+++ src/GovUk.Education.ExploreEducationStatistics.Data.Services/TimePeriodUtil.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
-using GovUk.Education.ExploreEducationStatistics.Data.Model.Query;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
 using GovUk.Education.ExploreEducationStatistics.Data.Services.Extensions;
 
 namespace GovUk.Education.ExploreEducationStatistics.Data.Services
