+++ src/GovUk.Education.ExploreEducationStatistics.Data.Services/ViewModels/Meta/ObservationalUnitGeoJsonMeta.cs
@@ -1,4 +1,5 @@
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data;
 using GovUk.Education.ExploreEducationStatistics.Data.Model;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
