+++ src/GovUk.Education.ExploreEducationStatistics.Data.Services/ViewModels/ObservationViewModel.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
+using GovUk.Education.ExploreEducationStatistics.Common.Model.Data;
 using GovUk.Education.ExploreEducationStatistics.Data.Model;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
